+++ screens/calendar.js
@@ -1,39 +1,7 @@
-import React from "react";
-import { View, Text, StyleSheet } from "react-native";
-import { useNavigation } from "@react-navigation/native";
+import React from 'react';
 
-const Calendar = () => {
-  const navigation = useNavigation();
-
-  return (
-    <View style={Styles.container}>
-      <Text style={Styles.HomeText}>캘린더 화면</Text>
-    </View>
-  );
+const calendar = () => {
+	return <div>calendar</div>;
 };
 
-export default Calendar;
-
-const Styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: "#fff",
-  },
-  HomeText: {
-    fontSize: 30,
-    textAlign: "center",
-  },
-  NextBottom: {
-    backgroundColor: "purple",
-    padding: 10,
-    marginTop: "20%",
-    width: "50%",
-    alignSelf: "center",
-    borderRadius: 10,
-  },
-  BottomText: {
-    fontSize: 15,
-    color: "white",
-    textAlign: "center",
-  },
-});
+export default calendar;
